Use less code
